# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A lightweight real-time chat backend written in Go. The server manages ephemeral chat rooms where clients connect via WebSocket to exchange messages. Rooms are created on-demand and automatically deleted after 3 hours of inactivity.\n\n## Additional Info\nMany resources (rooms, users, messages) carry an `additionalInfo` field. This is a free-form JSON object that the server stores as-is — it has no predefined schema and is never validated or interpreted by the backend.\n\nUse it to attach arbitrary metadata to any resource, for example:\n- **Rooms:** theme, description, language, or feature flags for your UI.\n- **Users:** avatar URL, bio, status text, or client-specific preferences.\n- **Messages:** formatting hints, link previews, or custom reaction data.\n\nThe field is always optional. If omitted it defaults to an empty object (`{}`). On `PATCH` requests the provided keys are merged into the existing object; on `PUT` the entire object is replaced.",
        "title": "Chat Room API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/rooms/{roomID}/messages": {
            "get": {
                "description": "Returns all messages that have been sent in a specific room. Messages are stored in memory and include system messages (joins/leaves) as well as user messages. Only messages smaller than 2 MiB are stored.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Get all messages in a room",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "roomID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MessagesListResponse"
                        }
                    },
                    "400": {
                        "description": "can't parse room id to uint",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "room not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/rooms/{roomID}/messages/{messageID}": {
            "get": {
                "description": "Retrieves a specific message from a room by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Get a specific message",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "roomID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Message UUID",
                        "name": "messageID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/OutgoingMessage"
                        }
                    },
                    "400": {
                        "description": "can't parse room id or message id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "room or message not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Completely replaces a message. Unlike PATCH, this requires all fields and replaces the entire message content. The server automatically sets modified: true in additionalInfo.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Replace a message",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "roomID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Message UUID",
                        "name": "messageID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New message content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MessagePutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/OutgoingMessage"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "room or message not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Marks a message as deleted. The message is not actually removed but its content is replaced with \"deleted\" and a deleted flag is added to additionalInfo.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Delete a message",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "roomID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Message UUID",
                        "name": "messageID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/OutgoingMessage"
                        }
                    },
                    "400": {
                        "description": "can't parse room id or message id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "room or message not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Partially updates a specific message. You can update the message text, additionalInfo, or both. Only provided fields are updated. The server automatically sets modified: true in additionalInfo.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Partially update a message",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "roomID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Message UUID",
                        "name": "messageID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MessagePatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/OutgoingMessage"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "room or message not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "MessageAdditionalInfo": {
            "type": "object",
            "properties": {
                "format": {
                    "type": "string",
                    "example": "markdown"
                },
                "modified": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "MessagePatchRequest": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "$ref": "#/definitions/MessageAdditionalInfo"
                },
                "message": {
                    "type": "string",
                    "example": "Hello everyone! (edited)"
                }
            }
        },
        "MessagePutRequest": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "$ref": "#/definitions/MessageAdditionalInfo"
                },
                "message": {
                    "type": "string",
                    "example": "Completely new message content"
                }
            }
        },
        "MessagesListResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/OutgoingMessage"
                    }
                }
            }
        },
        "OutgoingMessage": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "$ref": "#/definitions/MessageAdditionalInfo"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "message": {
                    "type": "string",
                    "example": "Hello everyone!"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-04-09T12:35:10.123456789Z"
                },
                "type": {
                    "type": "string",
                    "example": "message"
                },
                "user": {
                    "$ref": "#/definitions/User"
                }
            }
        },
...
        "User": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "$ref": "#/definitions/UserAdditionalInfo"
                },
                "firstName": {
                    "type": "string",
                    "example": "John"
                },
                "id": {
                    "type": "string",
                    "example": "9a6e58a5-4d47-4c86-8b3f-9ea373cbdb0c"
                },
                "lastName": {
                    "type": "string",
                    "example": "Doe"
                },
                "name": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
        "UserAdditionalInfo": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "https://example.com/avatar.png"
                },
                "bio": {
                    "type": "string",
                    "example": "Software developer"
                }
            }
        },
...
    }
}